split implementation for product, change package name
