+++ internal/product/injector/wire.go
@@ -1,23 +1,25 @@
 //go:build wireinject
 // +build wireinject
 
-package product
+package injector
 
 import (
 	handler "Dzaakk/simple-commerce/internal/product/handler"
 	repository "Dzaakk/simple-commerce/internal/product/repository"
+	routes "Dzaakk/simple-commerce/internal/product/routes"
 	usecase "Dzaakk/simple-commerce/internal/product/usecase"
 	"database/sql"
 
 	"github.com/google/wire"
 )
 
-func InitializedService(db *sql.DB) *handler.ProductHandler {
+func InitializedService(db *sql.DB) *routes.ProductRoutes {
 	wire.Build(
 		repository.NewProductRepository,
 		usecase.NewProductUseCase,
 		handler.NewProductHandler,
+		routes.NewProductRoutes,
 	)
 
-	return &handler.ProductHandler{}
+	return &routes.ProductRoutes{}
 }
