+++ internal/product/models/product.go
@@ -1,4 +1,4 @@
-package product
+package models
 
 import "Dzaakk/simple-commerce/package/template"
